Move `BlockingController` to `ui` package
